// src/fred_ts_client/models/Categories.ts
/**
 * FRED API
 * API for accessing economic data from the Federal Reserve Economic Data (FRED) database.
 *
 * OpenAPI spec version: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { CategoriesCategoriesInner } from '../models/CategoriesCategoriesInner';
import { HttpFile } from '../http/http';

export class Categories {
    'categories'?: Array<CategoriesCategoriesInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "categories",
            "baseName": "categories",
            "type": "Array<CategoriesCategoriesInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Categories.attributeTypeMap;
    }

    public constructor() {
    }
}